// types/schema/platform/app-domain/app-domain.types.ts
// //////////////////////////////////////////////////////////////////////////////
//  IMPORTS
// //////////////////////////////////////////////////////////////////////////////

// shared types
import * as B from '../../../shared/base';

// enums
import * as PlatformEnums from '../../../enums/platform';

// relations
import * as App from '../app/app.types';

// //////////////////////////////////////////////////////////////////////////////
//  CORE INTERFACE
// //////////////////////////////////////////////////////////////////////////////

export interface IAppDomain extends B.IBase {
  /*-------- Basic Info -------------*/
  domain: string;
  environment: PlatformEnums.EAppEnvironment;

  /*-------- Flags ------------------*/
  isPrimary: boolean;
  isActive: boolean;

  /*-------- Foreign Keys ----------*/
  appId: string;

  /*-------- Relations -------------*/
  readonly app?: App.IApp;
}

// //////////////////////////////////////////////////////////////////////////////
//  ROOT KEYS
// //////////////////////////////////////////////////////////////////////////////

export type AppDomainKey = keyof IAppDomain;